gdesear_CS446_PA4/requirements.txt
opencv-python
numpy
matplotlib
pyOpenSSL
